+++ papyrus/basepage.py
@@ -2,63 +2,39 @@ import frontmatter
 import pathlib
 import datetime
 import os
-import re
 
-class BasePage():
+class BasePage:
+
     def __init__(self, input_path):
         self.input_path = input_path
+        self.meta = None
+        self.date = None
+        self.title = None
+        self.props = None
+        self.layout = None
+        self.url = None
+        self.tags = None
+        self.output_path = None
         self.content_text = None
         self.content_html = None
         self.excerpt_text = None
         self.excerpt_html = None
-        self.output_path = None
-        self.layout = None
-        self.url = None
-        self.props = None
-        self.title = ""
-        self.date = None
-        self.tags = []
-
-        self.__update()
-
-    def __update(self):
-        with self.input_path.open() as infile:
-            content = infile.read()
-        # parse and remove front matter
-        front_meta, content = frontmatter.parse(content)
-        # parse and remove back matter
-        back_meta, content = self.parse_backmatter(content)
-
-        self.meta = front_meta | back_meta
-        self.title, self.content = self.extract_title(content)
-        if not self.title:
-            self.title = self.meta['title'] if 'title' in self.meta else None
-        self.layout = self.meta['layout'] if 'layout' in self.meta else 'post'
-        self.date = self.get_date(self.meta, self.input_path)
-        self.url = self.get_url(self.meta)
-        self.props = self.meta['props'].split(" ") if "props" in self.meta else []
-        self.tags = self.meta['tags'].split(" ") if ('tags' in self.meta and self.meta['tags']) else []
 
     def build(self, pandoc, config):
-        self.__update()
+        with self.input_path.open() as infile:
+            content = infile.read()
+        self.meta, content = frontmatter.parse(content)
 
-        self.excerpt_text, self.content_text = \
-            self.get_excerpt_content(self.content, config['EXCERPT_SEPARATOR'])
+        self.date = self.get_date(self.meta, self.input_path)
+        self.props = self.meta['props'].split(" ") if "props" in self.meta else []
+        self.title, content = self.extract_title(content)
+        self.layout = self.meta['layout'] if 'layout' in self.meta else 'post'
+        self.url = self.get_url(self.meta, self.input_path)
+        self.tags = self.meta['tags'].split(" ") if ('tags' in self.meta and self.meta['tags']) else []
+
         self.output_path = self.get_output_path(config['DIR_PUBLISH'])
-
-    @staticmethod
-    def parse_backmatter(content):
-        back_meta = {}
-        ii = []
-        for mo in re.finditer(r"^---\s*$", content, re.MULTILINE):
-            ii.append(mo.start())
-
-        if (len(ii)) == 2 and len(content) == ii[1] + 3:
-            back_matter = content[ii[0]:]
-            back_meta, _ = frontmatter.parse(back_matter)
-            content = content[:ii[0]].strip()
-
-        return back_meta, content
+        self.excerpt_text, self.content_text = \
+            self.get_excerpt_content(content, config['EXCERPT_SEPARATOR'])
 
     @staticmethod
     def extract_title(content):
@@ -87,7 +63,8 @@ class BasePage():
 
         return False
 
-    def get_date(self, meta, input_path):
+    @staticmethod
+    def get_date(meta, input_path):
         pagedate = None
         try:
             pagedate = datetime.date.fromisoformat(str(meta['date']))
@@ -95,7 +72,7 @@ class BasePage():
             pass
         if pagedate is None:
             try:
-                pagedate = datetime.date.fromisoformat(str(self.input_path.stem[:10]))
+                pagedate = datetime.date.fromisoformat(str(input_path.stem[:10]))
             except:
                 pass
         if pagedate is None:
@@ -103,14 +80,15 @@ class BasePage():
             pagedate = datetime.date.fromtimestamp(created_ts)
         return pagedate
 
-    def get_url(self, meta):
+    @staticmethod
+    def get_url(meta, input_path):
         if 'path' in meta:
             if meta['path'] == '/':
                 return "index.html"
             else:
                 return meta['path']
         else:
-            return self.input_path.with_suffix('').name
+            return input_path.with_suffix('').name
 
     def get_output_path(self, site_path):
         path = self.url
@@ -133,6 +111,7 @@ class BasePage():
         self.content_html = pandoc.markup(input=self.content_text)
         if self.excerpt_text is not None:
             self.excerpt_html = pandoc.markup(input=self.excerpt_text)
+            self.excerpt_html = self.excerpt_html.replace('<p>', '').replace('</p>', '')
 
     def is_current(self):
         if self.output_path is None or not self.output_path.exists():
@@ -142,13 +121,8 @@ class BasePage():
         else:
             return True
 
-    def render(self, jinja_env, site_meta, page_meta):
-        raise NotImplementedError()
-
     def write(self, jinja_env, site_meta):
-        template_out = self.render(jinja_env, site_meta, self.get_page_metadata())
-        self.output_path.parent.mkdir(parents=True, exist_ok=True)
-        self.output_path.write_text(template_out)
+        raise NotImplementedError("Method not implemented in base class")
 
     def get_page_metadata(self):
         page = {