+++ papyrus/clean.py
@@ -8,5 +8,3 @@ def clean(args, conf):
     if site_dir.exists():
         shutil.rmtree(site_dir)
     site_dir.mkdir()
-    post_store_path = pathlib.Path('.') / conf['POST_STORE']
-    post_store_path.unlink(missing_ok=True)